# public/swagger.json
{
	"components": {
...
		"schemas": {
...
			"Url": {
				"properties": {
					"type": {
						"type": "string"
					},
					"url": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"Image": {
				"properties": {
					"path": {
						"type": "string"
					},
					"extension": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"ComicSummary": {
				"properties": {
					"resourceURI": {
						"type": "string"
					},
					"name": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"ComicList": {
				"properties": {
					"available": {
						"type": "number",
						"format": "double"
					},
					"returned": {
						"type": "number",
						"format": "double"
					},
					"collectionURI": {
						"type": "string"
					},
					"items": {
						"items": {
							"$ref": "#/components/schemas/ComicSummary"
						},
						"type": "array"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"StorySummary": {
				"properties": {
					"resourceURI": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"type": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"StoryList": {
				"properties": {
					"available": {
						"type": "number",
						"format": "double"
					},
					"returned": {
						"type": "number",
						"format": "double"
					},
					"collectionURI": {
						"type": "string"
					},
					"items": {
						"items": {
							"$ref": "#/components/schemas/StorySummary"
						},
						"type": "array"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"EventSummary": {
				"properties": {
					"resourceURI": {
						"type": "string"
					},
					"name": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"EventList": {
				"properties": {
					"available": {
						"type": "number",
						"format": "double"
					},
					"returned": {
						"type": "number",
						"format": "double"
					},
					"collectionURI": {
						"type": "string"
					},
					"items": {
						"items": {
							"$ref": "#/components/schemas/EventSummary"
						},
						"type": "array"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"SeriesSummary": {
				"properties": {
					"resourceURI": {
						"type": "string"
					},
					"name": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"SeriesList": {
				"properties": {
					"available": {
						"type": "number",
						"format": "double"
					},
					"returned": {
						"type": "number",
						"format": "double"
					},
					"collectionURI": {
						"type": "string"
					},
					"items": {
						"items": {
							"$ref": "#/components/schemas/SeriesSummary"
						},
						"type": "array"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"Character": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"modified": {
						"type": "string",
						"format": "date-time"
					},
					"resourceURI": {
						"type": "string"
					},
					"urls": {
						"items": {
							"$ref": "#/components/schemas/Url"
						},
						"type": "array"
					},
					"thumbnail": {
						"$ref": "#/components/schemas/Image"
					},
					"comics": {
						"$ref": "#/components/schemas/ComicList"
					},
					"stories": {
						"$ref": "#/components/schemas/StoryList"
					},
					"events": {
						"$ref": "#/components/schemas/EventList"
					},
					"series": {
						"$ref": "#/components/schemas/SeriesList"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"CharacterDataContainer": {
				"properties": {
					"offset": {
						"type": "number",
						"format": "double"
					},
					"limit": {
						"type": "number",
						"format": "double"
					},
					"total": {
						"type": "number",
						"format": "double"
					},
					"count": {
						"type": "number",
						"format": "double"
					},
					"result": {
						"items": {
							"$ref": "#/components/schemas/Character"
						},
						"type": "array"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"CharacterDataWrapper": {
				"properties": {
					"code": {
						"type": "number",
						"format": "double"
					},
					"status": {
						"type": "string"
					},
					"copyright": {
						"type": "string"
					},
					"attributionText": {
						"type": "string"
					},
					"attributionHTML": {
						"type": "string"
					},
					"etag": {
						"type": "string"
					},
					"data": {
						"$ref": "#/components/schemas/CharacterDataContainer"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "marvel-xendit-api",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
...
		"/marvel/characters/{id}": {
			"get": {
				"operationId": "GetCharacterById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CharacterDataWrapper"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}